codes/ecg-image-digitizer/requirements.txt
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=0.24.0
Pillow>=8.0.0
matplotlib>=3.3.0
